extract common methods in headless-react
https://coveord.atlassian.net/browse/KIT-2769
