+++ packages/headless-react/src/ssr/index.ts
@@ -1,6 +1,4 @@
-export {
-  defineSearchEngine,
-  MissingEngineProviderError,
-} from './search-engine.js';
+export {defineSearchEngine} from './search-engine.js';
 export type {ReactSearchEngineDefinition} from './search-engine.js';
+export {MissingEngineProviderError} from './common.js';
 export * from '@coveo/headless/ssr';
